fixed TODOs
